refactor: move SmtpClient to it's own module
